requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
docopt==0.6.2
frozenlist==1.3.0
fsspec==2022.3.0
gcsfs==2022.3.0
google-api-core==2.7.3
google-auth==2.6.6
google-auth-oauthlib==0.5.1
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.1
idna==3.3
multidict==6.0.2
oauthlib==3.2.0
pandas==1.4.2
pipreqs==0.4.11
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
yarg==0.1.9
yarl==1.7.2